refactor: split app_cmd.py (1,416 lines) into focused modules (fixes #832)
Moved all code from naturo/cli/app_cmd.py into naturo/cli/_app/ package:
- _common.py: shared helpers (app ID resolution, error formatting)
- lifecycle.py: launch, quit, relaunch, list, find (370 lines)
- diagnostics.py: inspect + detection scanning (290 lines)
- window_ops.py: focus, close, minimize, maximize, restore, move (420 lines)
- legacy.py: hide, unhide, switch hidden aliases (160 lines)

app_cmd.py remains as a thin re-export shim so all existing imports
(tests, CLI registration, cross-module references) continue to work.
All 4652 tests pass, ruff/mypy clean.
